corrigiendo ejercicios del tema 3

+++ Ejercicios Java Tema 3/E2t3.java
@@ -8,7 +8,7 @@ public class E2t3 { // Clase principal
   System.out.println("Por favor introduce un numero en euros: ");
   double euros = Integer.parseInt( System.console().readLine());
   
-  double pesetas = ((int) euros * (int) 166.386);
+  int pesetas = (int)(euros * 166.386);
   System.out.println(pesetas + " Son las pesetas que valen "+ euros +" euros");
 
   }
